Fix for SA-207 and SA-209
