+++ ds3-autogen-net/src/main/resources/tmpls/net/request/bulk_get_request.ftl
@@ -1,12 +1,11 @@
 <#include "../common/copyright.ftl" />
 
+using Ds3.Calls.Util;
 using Ds3.Models;
-using Ds3.Runtime;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Xml.Linq;
 
 namespace Ds3.Calls
 {
@@ -39,8 +38,8 @@ namespace Ds3.Calls
 
         <#include "common/constructor.ftl" />
 
-        public ${name}(string bucketName, List<Ds3Object> objects)
-            : this(bucketName, objects.Select(o => o.Name), Enumerable.Empty<Ds3PartialObject>())
+        public ${name}(string bucketName, List<Ds3Object> ds3Objects)
+            : this(bucketName, ds3Objects.Select(o => o.Name), Enumerable.Empty<Ds3PartialObject>())
         {
         }
 
@@ -48,19 +47,7 @@ namespace Ds3.Calls
 
         internal override Stream GetContentStream()
         {
-            var root = new XElement("Objects")
-                .AddAllFluent(
-                    from name in this.FullObjects
-                    select new XElement("Object").SetAttributeValueFluent("Name", name)
-                )
-                .AddAllFluent(
-                    from partial in this.PartialObjects
-                    select new XElement("Object")
-                        .SetAttributeValueFluent("Name", partial.Name)
-                        .SetAttributeValueFluent("Offset", partial.Range.Start.ToString())
-                        .SetAttributeValueFluent("Length", partial.Range.Length.ToString())
-                );
-            return new XDocument().AddFluent(root).WriteToMemoryStream();
+            return RequestPayloadUtil.MarshalFullAndPartialObjects(this.FullObjects, this.PartialObjects);
         }
 
         internal override long GetContentLength()
